backend/requirements.txt
# FastAPI y dependencias core
fastapi==0.115.13
uvicorn==0.34.3
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.1.0

# Seguridad y Authentication
fastapi-users[sqlalchemy]==14.0.1 

# Agentes AI - VERSIONES ESTABLES (COMPATIBLES)
# 1.89.1: latest stable line. Provides AbstractCapability, Hooks, ReinjectSystemPrompt,
# UsageLimits, output_validator, BedrockConverseModel + prompt caching.
# Extra is [bedrock] (pulls boto3 metadata used by the Bedrock provider). The OpenAI
# SDK is brought in independently below for non-pydantic-ai usage (Whisper, etc.).
pydantic-ai-slim[bedrock]==1.89.1
openai>=1.75.0
python-multipart==0.0.20

# Base de datos
sqlalchemy==2.0.41
alembic==1.16.2
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Redis
redis==6.4.0  # Latest - includes redis.asyncio

# HTTP y Networking
httpx==0.28.1
httpcore==1.0.9
requests==2.32.4

# AWS 
# S3 access uses sync boto3 wrapped in asyncio.to_thread (see app/services/s3_service.py).
# aioboto3/aiobotocore were dropped: they crónicamente atrasan el floor de boto3 vs
# pydantic-ai-slim, lo que bloqueaba el upgrade del agente.
boto3>=1.42.63
botocore>=1.42.63

# Procesamiento de Documentos - CRÍTICO PARA TUS PDFs
markdown==3.8.2
jinja2==3.1.6
aiofiles==24.1.0
openpyxl==3.1.5
python-docx==1.1.2
PyPDF2==3.0.1

# Visualizaciones - CRÍTICO PARA TUS CHARTS
matplotlib==3.10.3
numpy==2.3.0
plotly==5.17.0
kaleido==0.2.1
pydot==2.0.0
pillow==11.2.1

weasyprint==66.0  # Updated for Phase 1: PDF Generation
tinycss2==1.4.0
cssselect2==0.8.0
pydyf==0.11.0
cffi==1.17.1
Brotli==1.1.0

# Utilidades
PyYAML==6.0.2
click==8.2.1
rich==14.0.0
typer==0.16.0
tenacity==9.1.2

slowapi==0.1.9  # Rate limiting for authentication endpoints
aiosmtplib

structlog==24.4.0  # Modern structured logging with JSON output

gunicorn==22.0.0

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
